+++ src/types/zeta.ts
@@ -1071,6 +1071,27 @@ export type Zeta = {
       "args": []
     },
     {
+      "name": "updateMakerTradeFeePercentage",
+      "accounts": [
+        {
+          "name": "state",
+          "isMut": true,
+          "isSigner": false
+        },
+        {
+          "name": "admin",
+          "isMut": false,
+          "isSigner": true
+        }
+      ],
+      "args": [
+        {
+          "name": "newNativeMakerTradeFeePercentage",
+          "type": "u64"
+        }
+      ]
+    },
+    {
       "name": "updateZetaState",
       "accounts": [
         {
@@ -6333,6 +6354,34 @@ export type Zeta = {
       ]
     },
     {
+      "name": "editMaType",
+      "accounts": [
+        {
+          "name": "state",
+          "isMut": false,
+          "isSigner": false
+        },
+        {
+          "name": "admin",
+          "isMut": false,
+          "isSigner": true
+        },
+        {
+          "name": "marginAccount",
+          "isMut": true,
+          "isSigner": false
+        }
+      ],
+      "args": [
+        {
+          "name": "maType",
+          "type": {
+            "defined": "MarginAccountType"
+          }
+        }
+      ]
+    },
+    {
       "name": "initializeReferrerAccount",
       "accounts": [
         {
@@ -7249,11 +7298,15 @@ export type Zeta = {
             }
           },
           {
+            "name": "nativeMakerTradeFeePercentage",
+            "type": "u64"
+          },
+          {
             "name": "padding",
             "type": {
               "array": [
                 "u8",
-                106
+                98
               ]
             }
           }
@@ -9428,6 +9481,9 @@ export type Zeta = {
           },
           {
             "name": "MarketMaker"
+          },
+          {
+            "name": "MarketMakerT1"
           }
         ]
       }
@@ -11774,6 +11830,27 @@ export const IDL: Zeta = {
       "args": []
     },
     {
+      "name": "updateMakerTradeFeePercentage",
+      "accounts": [
+        {
+          "name": "state",
+          "isMut": true,
+          "isSigner": false
+        },
+        {
+          "name": "admin",
+          "isMut": false,
+          "isSigner": true
+        }
+      ],
+      "args": [
+        {
+          "name": "newNativeMakerTradeFeePercentage",
+          "type": "u64"
+        }
+      ]
+    },
+    {
       "name": "updateZetaState",
       "accounts": [
         {
@@ -17036,6 +17113,34 @@ export const IDL: Zeta = {
       ]
     },
     {
+      "name": "editMaType",
+      "accounts": [
+        {
+          "name": "state",
+          "isMut": false,
+          "isSigner": false
+        },
+        {
+          "name": "admin",
+          "isMut": false,
+          "isSigner": true
+        },
+        {
+          "name": "marginAccount",
+          "isMut": true,
+          "isSigner": false
+        }
+      ],
+      "args": [
+        {
+          "name": "maType",
+          "type": {
+            "defined": "MarginAccountType"
+          }
+        }
+      ]
+    },
+    {
       "name": "initializeReferrerAccount",
       "accounts": [
         {
@@ -17952,11 +18057,15 @@ export const IDL: Zeta = {
             }
           },
           {
+            "name": "nativeMakerTradeFeePercentage",
+            "type": "u64"
+          },
+          {
             "name": "padding",
             "type": {
               "array": [
                 "u8",
-                106
+                98
               ]
             }
           }
@@ -20131,6 +20240,9 @@ export const IDL: Zeta = {
           },
           {
             "name": "MarketMaker"
+          },
+          {
+            "name": "MarketMakerT1"
           }
         ]
       }